Move _drop_rows_by_label from Frame to IndexedFrame (#11157)
This PR is a trivial move of a function from frame.py to indexed_frame.py. That function is only used in one place throughout cuDF, inside `IndexedFrame.drop`.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - https://github.com/brandon-b-miller
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/11157
